--- src/devices_models/devices/common.jl
@@ -1,14 +1,0 @@
-include("common/constraints_structs.jl")
-include("common/cost_functions.jl")
-include("common/range_constraint.jl")
-include("common/add_variable.jl")
-include("common/add_parameters.jl")
-include("common/rating_constraints.jl")
-include("common/rateofchange_constraints.jl")
-include("common/duration_constraints.jl")
-include("common/commitment_constraint.jl")
-include("common/timeseries_constraint.jl")
-include("common/expressionarray_algebra.jl")
-include("common/pm_translator.jl")
-include("common/energy_balance_constraint.jl")
-include("common/get_time_series.jl")
