--- src/services_models/services_model.jl
@@ -1,6 +1,0 @@
-abstract type AbstractServiceFormulation end
-
-mutable struct ServiceModel{D <: PSY.Service, B <: AbstractServiceFormulation}
-    service_type::Type{D}
-    formulation::Type{B}
-end
